+++ Forge/src/main/java/fuzs/puzzleslib/impl/client/core/context/BuildCreativeModeTabContentsContextForgeImpl.java
@@ -1,29 +1,19 @@
 package fuzs.puzzleslib.impl.client.core.context;
 
 import fuzs.puzzleslib.api.client.core.v1.context.BuildCreativeModeTabContentsContext;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.item.CreativeModeTab;
-import net.minecraftforge.common.CreativeModeTabRegistry;
 
 import java.util.Objects;
 
-public record BuildCreativeModeTabContentsContextForgeImpl(ResourceLocation identifier,
+public record BuildCreativeModeTabContentsContextForgeImpl(ResourceKey<CreativeModeTab> resourceKey,
                                                            CreativeModeTab.ItemDisplayParameters parameters,
                                                            CreativeModeTab.Output output) implements BuildCreativeModeTabContentsContext {
 
     @Override
-    public void registerBuildListener(ResourceLocation identifier, CreativeModeTab.DisplayItemsGenerator itemsGenerator) {
-        Objects.requireNonNull(identifier, "identifier is null");
+    public void registerBuildListener(ResourceKey<CreativeModeTab> resourceKey, CreativeModeTab.DisplayItemsGenerator itemsGenerator) {
+        Objects.requireNonNull(resourceKey, "resource key is null");
         Objects.requireNonNull(itemsGenerator, "display items generator is null");
-        if (Objects.equals(identifier, this.identifier)) {
-            itemsGenerator.accept(this.parameters, this.output);
-        }
-    }
-
-    @Override
-    public void registerBuildListener(CreativeModeTab tab, CreativeModeTab.DisplayItemsGenerator itemsGenerator) {
-        Objects.requireNonNull(tab, "creative mode tab is null");
-        ResourceLocation identifier = CreativeModeTabRegistry.getName(tab);
-        this.registerBuildListener(identifier, itemsGenerator);
+        if (resourceKey == this.resourceKey) itemsGenerator.accept(this.parameters, this.output);
     }
 }